common: Move GetHostDistroName to common
So that it can be used by composer-cli cmd functions like listing the
image types.

Related: RHEL-60134
