+++ internal/common/osrelease_test.go
@@ -1,4 +1,4 @@
-package cloud
+package common
 
 import (
 	"strings"
